# api/docs/swagger.yaml
basePath: /local-eats
definitions:
  dish.DishDetails:
    properties:
      available:
        type: boolean
      category:
        type: string
      id:
        type: string
      name:
        type: string
      price:
        type: number
    type: object
  dish.DishInfo:
    properties:
      allergens:
        items:
          type: string
        type: array
      available:
        type: boolean
      category:
        type: string
      created_at:
        type: string
      description:
        type: string
      dietary_info:
        items:
          type: string
        type: array
      id:
        type: string
      ingredients:
        items:
          type: string
        type: array
      kitchen_id:
        type: string
      name:
        type: string
      nutrition_info:
        $ref: '#/definitions/dish.NutritionalInfo'
      price:
        type: number
      updated_at:
        type: string
    type: object
  dish.Dishes:
    properties:
      dishes:
        items:
          $ref: '#/definitions/dish.DishDetails'
        type: array
      limit:
        type: integer
      page:
        type: integer
      total:
        type: integer
    type: object
  dish.NewDataNoID:
    properties:
      available:
        type: boolean
      name:
        type: string
      price:
        type: number
    type: object
  dish.NewDish:
    properties:
      available:
        type: boolean
      category:
        type: string
      description:
        type: string
      ingredients:
        items:
          type: string
        type: array
      kitchen_id:
        type: string
      name:
        type: string
      price:
        type: number
    type: object
  dish.NewDishResp:
    properties:
      available:
        type: boolean
      category:
        type: string
      created_at:
        type: string
      description:
        type: string
      id:
        type: string
      ingredients:
        items:
          type: string
        type: array
      kitchen_id:
        type: string
      name:
        type: string
      price:
        type: number
    type: object
  dish.NutritionalInfo:
    properties:
      calories:
        type: integer
      carbs:
        type: integer
      fat:
        type: integer
      protein:
        type: integer
    type: object
  dish.UpdatedData:
    properties:
      available:
        type: boolean
      category:
        type: string
      description:
        type: string
      id:
        type: string
      ingredients:
        items:
          type: string
        type: array
      kitchen_id:
        type: string
      name:
        type: string
      price:
        type: number
      updated_at:
        type: string
    type: object
  extra.Activity:
    properties:
      favorite_cuisines:
        items:
          $ref: '#/definitions/extra.Cuisine'
        type: array
      favorite_kitchens:
        items:
          $ref: '#/definitions/extra.Kitchen'
        type: array
      total_orders:
        type: integer
      total_spent:
        type: number
    type: object
  extra.Cuisine:
    properties:
      cuisine_type:
        type: string
      orders_count:
        type: integer
    type: object
  extra.DaySchedule:
    properties:
      close:
        type: string
      open:
        type: string
    type: object
  extra.Dish:
    properties:
      id:
        type: string
      name:
        type: string
      orders_count:
        type: integer
      revenue:
        type: number
    type: object
  extra.Kitchen:
    properties:
      id:
        type: string
      name:
        type: string
      orders_count:
        type: integer
    type: object
  extra.NutritionalInfo:
    properties:
      allergens:
        items:
          type: string
        type: array
      calories:
        type: integer
      carbs:
        type: integer
      dietary_info:
        items:
          type: string
        type: array
      fat:
        type: integer
      protein:
        type: integer
    type: object
  extra.Statistics:
    properties:
      average_rating:
        type: number
      top_dishes:
        items:
          $ref: '#/definitions/extra.Dish'
        type: array
      total_orders:
        type: integer
      total_revenue:
        type: number
    type: object
  extra.WorkingHoursResp:
    properties:
      kitchen_id:
        type: string
      schedule:
        additionalProperties:
          $ref: '#/definitions/extra.DaySchedule'
        type: object
      updated_at:
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: API Gateway of Local Eats
  title: Local Eats
  version: "1.0"
paths:
  /dishes:
    post:
      description: Inserts a new dish into database
      parameters:
      - description: Dish info
        in: body
        name: dish
        required: true
        schema:
          $ref: '#/definitions/dish.NewDish'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dish.NewDishResp'
        "400":
          description: Invalid dish data
          schema:
            type: string
        "500":
          description: Server error while processing request
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Creates a dish
      tags:
      - dish
  /dishes/{id}:
    delete:
      description: Deletes dish from database
      parameters:
      - description: Dish ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Invalid dish ID
          schema:
            type: string
        "500":
          description: Server error while processing request
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Deletes a dish
      tags:
      - dish
    get:
      description: Retrieves dish info from database
      parameters:
      - description: Dish ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dish.DishInfo'
        "400":
          description: Invalid dish ID
          schema:
            type: string
        "500":
          description: Server error while processing request
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Gets a dish
      tags:
      - dish
    put:
      description: Updates dish info in database
      parameters:
      - description: Dish ID
        in: path
        name: id
        required: true
        type: string
      - description: Dish info
        in: body
        name: dish
        required: true
        schema:
          $ref: '#/definitions/dish.NewDataNoID'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dish.UpdatedData'
        "400":
          description: Invalid dish ID
          schema:
            type: string
        "500":
          description: Server error while processing request
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Updates a dish
      tags:
      - dish
  /dishes/{id}/nutrition:
    get:
      description: Informs about dish's nutritional value
      parameters:
      - description: Dish ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/extra.NutritionalInfo'
        "400":
          description: Invalid dish ID
          schema:
            type: string
        "500":
          description: Server error while processing request
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Gets dish's nutrition info
      tags:
      - dish
...
  /kitchens/{id}/dishes:
    get:
      description: Retrieves dishes info from database
      parameters:
      - description: Page number
        in: query
        name: page
        required: true
        type: integer
      - description: Number of items per page
        in: query
        name: limit
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dish.Dishes'
        "500":
          description: Server error while processing request
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Gets dishes
      tags:
      - kitchen
...
  /kitchens/{id}/statistics:
    get:
      description: Informs about kitchen statistics by date
      parameters:
      - description: Kitchen ID
        in: path
        name: id
        required: true
        type: string
      - description: start date
        in: query
        name: start_date
        required: true
        type: string
      - description: end date
        in: query
        name: end_date
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/extra.Statistics'
        "400":
          description: Invalid kitchen ID or date format
          schema:
            type: string
        "500":
          description: Server error while processing request
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Gets kitchen's statistics
      tags:
      - kitchen
  /kitchens/{id}/working-hours:
    post:
      description: Sets working hours for kitchen
      parameters:
      - description: Kitchen ID
        in: path
        name: id
        required: true
        type: string
      - description: Working hours
        in: body
        name: schedule
        required: true
        schema:
          additionalProperties:
            $ref: '#/definitions/extra.DaySchedule'
          type: object
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/extra.WorkingHoursResp'
        "400":
          description: Invalid kitchen ID or data
          schema:
            type: string
        "500":
          description: Server error while processing request
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Sets working hours
      tags:
      - kitchen
...
  /users/{id}/activity:
    get:
      description: Informs about user's activity by date
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      - description: start date
        in: query
        name: start_date
        required: true
        type: string
      - description: end date
        in: query
        name: end_date
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/extra.Activity'
        "400":
          description: Invalid user ID or date format
          schema:
            type: string
        "500":
          description: Server error while processing request
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Tracks user's activity
      tags:
      - user
schemes:
- http
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"